Updated CameraSubscriber;
