+++ image_transport/CMakeLists.txt
@@ -25,7 +25,6 @@ add_library(${PROJECT_NAME}
   src/subscriber.cpp
   src/single_subscriber_publisher.cpp
   src/camera_publisher.cpp
-  src/camera_subscriber.cpp
   src/image_transport.cpp
 )
 target_include_directories(${PROJECT_NAME} PUBLIC
