+++ image_transport/include/image_transport/subscriber_filter.hpp
@@ -36,7 +36,7 @@
 #include "sensor_msgs/msg/image.hpp"
 #include "message_filters/simple_filter.h"
 
-#include "image_transport/image_transport.hpp"
+#include "image_transport/create_subscription.hpp"
 #include "image_transport/visibility_control.hpp"
 
 namespace image_transport
@@ -62,6 +62,7 @@ void callback(const std::shared_ptr<const sensor_msgs::msg::Image>&);
 \endverbatim
  */
 
+template<class NodeT = rclcpp::Node>
 class SubscriberFilter : public message_filters::SimpleFilter<sensor_msgs::msg::Image>
 {
 public:
@@ -75,9 +76,8 @@ public:
    * \param queue_size The subscription queue size
    * \param transport The transport hint to pass along
    */
-  IMAGE_TRANSPORT_PUBLIC
   SubscriberFilter(
-    rclcpp::Node * node, const std::string & base_topic,
+    NodeT * node, const std::string & base_topic,
     const std::string & transport)
   {
     subscribe(node, base_topic, transport);
@@ -107,7 +107,7 @@ public:
    */
   IMAGE_TRANSPORT_PUBLIC
   void subscribe(
-    rclcpp::Node * node,
+    NodeT * node,
     const std::string & base_topic,
     const std::string & transport,
     rmw_qos_profile_t custom_qos = rmw_qos_profile_default,
