+++ image_transport/test/test_message_passing.cpp
@@ -34,6 +34,9 @@
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp/node.hpp"
 
+#include "image_transport/create_publisher.hpp"
+#include "image_transport/create_camera_subscription.hpp"
+#include "image_transport/create_camera_publisher.hpp"
 #include "image_transport/image_transport.hpp"
 #include "sensor_msgs/msg/image.hpp"
