+++ image_transport/test/test_publisher.cpp
@@ -33,6 +33,8 @@
 
 #include "rclcpp/rclcpp.hpp"
 
+#include "image_transport/create_publisher.hpp"
+#include "image_transport/create_camera_publisher.hpp"
 #include "image_transport/image_transport.hpp"
 
 class TestPublisher : public ::testing::Test
